Add: ✨ wp_localize_script and nonce
It's will always guarantee the base url is always correct
if we are going to one day work with the Script based on staging or
PROD server when sending POST http request.

Nonce is something like security hash to prevent attacks and mistakes.
It creates unique identifiers to check if the ajax request is coming
from website or not.

In WordPress words A nonce is a "number used once" to help protect URLs
and forms from certain types of misuse, malicious or otherwise.
WordPress nonces aren't numbers, but are a hash made up of numbers and letters.
Nor are they used only once, but have a limited "lifetime" after which they expire.

Source:
https://wordpress.stackexchange.com/questions/231797/what-is-nonce-and-how-to-use-it-with-ajax-in-wordpress