Add: ✨ Saving the ACF JSON locally
Saving the ACF JSON files locally allows you to have version control
over your field settings! Let’s learn how to do that.

Also, it's worth to now that there is a file called:

back-up-acf-export-form-tax-vat-calculation_fields.json

This file will serve as a backup just in case if the ACF storing the local
file doesn't work. We can always use that file to import it to the WP
system